+++ Source/Private/ASDisplayNodeInternal.h
@@ -29,6 +29,9 @@
 #import <AsyncDisplayKit/_ASTransitionContext.h>
 #import <AsyncDisplayKit/ASWeakSet.h>
 
+#define _pendingDisplayNodeLayout ASLayoutElementContextGetPendingLayout(self)
+#define _pendingDisplayNodeLayoutSet(x) ASLayoutElementContextSetPendingLayout(self, x)
+
 NS_ASSUME_NONNULL_BEGIN
 
 @protocol _ASDisplayLayerDelegate;
@@ -158,7 +161,6 @@ FOUNDATION_EXPORT NSString * const ASRenderingEngineDidDisplayNodesScheduledBefo
   std::atomic<int32_t> _pendingTransitionID;
   ASLayoutTransition *_pendingLayoutTransition;
   std::shared_ptr<ASDisplayNodeLayout> _calculatedDisplayNodeLayout;
-  std::shared_ptr<ASDisplayNodeLayout> _pendingDisplayNodeLayout;
   
   ASDisplayNodeViewBlock _viewBlock;
   ASDisplayNodeLayerBlock _layerBlock;